src/requirements.txt
certifi
chardet
click
Flask
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
marshmallow
PyYAML
requests
six
urllib3
Werkzeug
flask-sqlalchemy
sqlalchemy-utils
flask-migrate
pypandoc
